django/backend/requirements.txt
firebase-admin
django-cors-headers
djangorestframework-simplejwt
psycopg2
psycopg2-binary
boto3
python-dotenv
Django
drf_yasg
Pillow
folium
gunicorn
